% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_package.R
\name{add_to_package}
\alias{add_to_package}
\title{Add content to package}
\usage{
add_to_package(package, content)
}
\arguments{
\item{package}{An object with class ContentPackage.}

\item{content}{An object with class Content.}
}
\value{
A package with content added.
}
\description{
Add content to package
}
\examples{
x <- create_content_package() |>
 add_to_package(
  create_content(
   create_content_backbone(
    'analysis of dataset',
    'T',
    function(dataset) {
     dataset |>
      flextable::flextable()
    }
   ),
   dataset = tibble::tibble(x = c(1, 2, 3))
  )
)
}
