% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_derive_numcat.R
\name{auto_derive_numcat}
\alias{auto_derive_numcat}
\title{Automatically Derive Numeric Categories}
\usage{
auto_derive_numcat(dataset, numcat_method, by)
}
\arguments{
\item{dataset}{A data frame or tibble. Columns must have `method_id`, `origin`, and `label` attributes.}

\item{numcat_method}{A string specifying the `method_id` to filter variables for derivation.}

\item{by}{Symbol. The subject identifier variable name common to `.data` and all datasets in `from`. Used to match rows across datasets (default is `"USUBJID"`).}
}
\value{
The modified dataset with derived numeric categories applied where applicable.
}
\description{
Automates the derivation of numeric categories in a dataset by identifying variable pairs based on their attributes and applying the `derive2` function.
}
\details{
Filters variables with `method_id == numcat_method` and `origin == "Derived"`. Identifies pairs where one variable's `label` ends with "(N)" and another has the same `label` without "(N)`. Applies `derive2` to such pairs. Lists variables that could not be derived automatically.
}
\examples{
# Example usage
auto_derive(dataset, "example_method", by = rlang::exprs(USUBJID))

}
