% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.R
\name{build_report}
\alias{build_report}
\title{Build Report Package for Submission}
\usage{
build_report(pkg_name)
}
\arguments{
\item{pkg_name}{A character string specifying the name of the package.}
}
\value{
The path to the generated folder.
}
\description{
This function generates a comprehensive package with log file for a specified report package and the rtfs associated. The log includes session information, backbone file details, and caller content logs. The package name must be provided as a character string, and the corresponding directory must exist in '03_Algorithm'.
}
\examples{
\dontrun{
# Example usage:
# Build a log file for the package 'Teste'
x <- build_report('Teste')
}
}
