% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_submission.R
\name{build_submission}
\alias{build_submission}
\title{Build a file with log for submission}
\usage{
build_submission(pkg_name, dataset = FALSE)
}
\arguments{
\item{pkg_name}{A character to specify the name of the package.}

\item{dataset}{A boolean to specify if the package is a dataset package, not a report.}
}
\value{
path to the log file.
}
\description{
Build a file with log for submission
}
\examples{
x <- build_submission('Teste')
}
