% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adam_datasets.R
\name{create_adam_datasets}
\alias{create_adam_datasets}
\title{Create Analysis Datasets}
\usage{
create_adam_datasets(path)
}
\arguments{
\item{path}{A character string specifying the path to the metadata file in .xlsx format.}
}
\value{
A list of tibbles, each representing a dataset defined in the metadata file, with zero rows.
}
\description{
This function generates an analysis dataset structure with zero rows based on the provided metadata file in .xlsx format. The metadata should follow the package template, including sheets for datasets, variables, and codelists.
}
\examples{
\dontrun{
# Example usage:
# Create an analysis datasets structure from 'Spec.xlsx'
x <- create_adam_datasets('Spec.xlsx')
}
}
