% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_backbone_structure.R
\name{create_backbone_structure}
\alias{create_backbone_structure}
\title{Create Backbone Structure Folder}
\usage{
create_backbone_structure(backbone_name, backbone_type, pkg_name)
}
\arguments{
\item{backbone_name}{A character string specifying the name of the backbone.}

\item{backbone_type}{A character string specifying the type of the backbone. Please use 'T', 'F', or 'L'.}

\item{pkg_name}{A character string specifying the name of the package.}
}
\value{
The path to the created backbone structure folder.
}
\description{
This function creates a backbone structure folder for a specified package. It requires the name of the backbone, the type of the backbone ('T' for table, 'F' for figure, or 'L' for listing), and the name of the package.
}
\examples{
\dontrun{
# Example usage:
# Create a backbone structure for a table named 'teste3' in package 'teste2'
create_backbone_structure('teste3', 'T', 'teste2')
}
}
