% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_content.R
\name{create_content}
\alias{create_content}
\title{Create Content from Content Backbone}
\usage{
create_content(
  content_backbone,
  subtitle = NA_character_,
  population = NA_character_,
  section = NA_character_,
  fdim = list(width = 9, height = 5, dpi = 600),
  export_name = NA_character_,
  ...
)
}
\arguments{
\item{content_backbone}{An object of class `ContentBackbone`.}

\item{subtitle}{A character string specifying the content subtitle (optional).}

\item{population}{A character string specifying the content population (optional).}

\item{section}{A character string specifying the content section (optional).}

\item{fdim}{A list of 3 elements: 'width', 'height', and 'dpi', each with a numeric value. This is only used when the type is 'F' (figure). Max Width, Height: 9 and 5. Min dpi: 600.}

\item{export_name}{A character string specifying the content export name. Must be unique between contents. If not specified, then a default name to content will be generated "adyyi". Only used for datasets export.}

\item{...}{Extra arguments to pass to the 'fun' in the content backbone (optional).}
}
\value{
A content object.
}
\description{
This function creates content using a specified content backbone and dataset. Additional parameters allow customization of subtitles, population, section, and dimensions for figures.
}
\examples{
\dontrun{
# Example usage:
# Create content from a content backbone with a dataset
x <- create_content(
 create_content_backbone(
  'analysis of dataset',
  'T',
  function(dataset) {
   dataset |>
    flextable::flextable()
  }
 ),
 dataset = tibble::tibble(x = c(1, 2, 3))
)
}
}
