% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_content_backbone.R
\name{create_content_backbone}
\alias{create_content_backbone}
\title{Create content backbone}
\usage{
create_content_backbone(title, type, fun)
}
\arguments{
\item{title}{A character to specify the title of the content.}

\item{type}{A character to specify the type of the content. Use 'T', 'F' or 'L'.}

\item{fun}{A function that creates the content. The return of the function must be a flextable for 'T', ggplot for 'F' and tibble for 'L'.}
}
\value{
A content backbone.
}
\description{
Create content backbone
}
\examples{
x <- create_content_backbone(
 'analysis of dataset',
 'T',
 function(dataset) {
  dataset |> flextable::flextable()
 }
)
}
