% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets_structure.R
\name{create_datasets_structure}
\alias{create_datasets_structure}
\title{Create Package Structure Folder}
\usage{
create_datasets_structure(pkg_name)
}
\arguments{
\item{pkg_name}{A character string specifying the name of the package.}
}
\value{
The path to the created package structure folder.
}
\description{
This function creates a package structure folder for a specified package name within the '03_Algorithm' directory. It sets up necessary template files and directories for the package.
}
\examples{
\dontrun{
# Example usage:
# Create a package structure for 'Teste'
create_datasets_structure('Teste')
}
}
