% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create project with template}
\usage{
create_project(proj_name, path = NA_character_, open = TRUE, rstudio = TRUE)
}
\arguments{
\item{proj_name}{A character to specify the name of the project.}

\item{path}{A character to specify the path to create the project.}

\item{open}{A boolean to specify whether the project should be opened.}

\item{rstudio}{A boolean to specify whether the rstudio project file should be created.}
}
\value{
A path to the project.
}
\description{
Create project with template
}
\examples{
create_project('Teste')
}
