% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report_structure.R
\name{create_report_structure}
\alias{create_report_structure}
\title{Create Report Structure Folder}
\usage{
create_report_structure(pkg_name)
}
\arguments{
\item{pkg_name}{A character string specifying the name of the package.}
}
\value{
The path to the created report structure folder.
}
\description{
This function creates a report structure folder for a specified package name within the '03_Algorithm' directory. It sets up necessary template files and directories for the report.
}
\examples{
\dontrun{
# Example usage:
# Create a report structure for 'Teste'
create_report_structure('Teste')
}
}
