% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_to_html.R
\name{define_to_html}
\alias{define_to_html}
\title{Convert CDISC define.xml to HTML using XSLT}
\usage{
define_to_html(xml_file, xsl_file, output_file = "define.html")
}
\arguments{
\item{xml_file}{Path to the CDISC define.xml file.}

\item{xsl_file}{Path to the XSL file to use for transformation.}

\item{output_file}{Path to the output HTML file. Defaults to "define.html".}
}
\value{
Invisibly returns the XML object of the transformed HTML.
}
\description{
Reads a CDISC define.xml file and an XSL file, applies the XSL transformation,
and saves the resulting HTML file.
}
\examples{
\dontrun{
define_to_html(
  xml_file = "path/to/define.xml",
  xsl_file = "path/to/define2-1-0.xsl",
  output_file = "define.html"
)
}
}
