% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive2.R
\name{derive2}
\alias{derive2}
\title{Simplified version of derive function}
\usage{
derive2(dataset, var_target, var_source, by = rlang::exprs(USUBJID))
}
\arguments{
\item{dataset}{A data frame (primary dataset) where the variable will be derived}

\item{var_target}{Symbol. Name of the variable to be derived in `dataset`.}

\item{var_source}{Symbol. Name of the variable to be used as reference to derive in `dataset`.}

\item{by}{Symbol. The subject identifier variable name common to `dataset`.}
}
\value{
The original dataset `dataset` with the variable `var_target` derived accordingly.
}
\description{
Simplified version of derive function to be used in situations where want to convert
one variable into another with same length of levels.
}
\examples{
\dontrun{
adsl <- deriveN(adsl, SEX, SEXN, by = USUBJID)
}

}
