% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_datasets.R
\name{export_datasets}
\alias{export_datasets}
\title{Export Dataset Package}
\usage{
export_datasets(package, spec)
}
\arguments{
\item{package}{An object of class `ContentPackage`.}

\item{spec}{A character string with path to the spec file.}
}
\value{
The path to the exported package.
}
\description{
This function exports a dataset package to a specified file format.
}
\examples{
\dontrun{
# Example usage:
# Create and export a content package
x <- create_content_package() |>
 add_to_package(
  create_content(
   create_content_backbone(
    'analysis of dataset',
    'T',
    function(dataset) {
     dataset |>
      flextable::flextable()
    }
   ),
   dataset = tibble::tibble(x = c(1, 2, 3))
  )
) |>
export_package('teste')
}
}
