% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_package.R
\name{export_package}
\alias{export_package}
\title{Export package}
\usage{
export_package(
  package,
  report_name,
  template_name = "template_PT-BR.docx",
  supp = FALSE,
  dataset = FALSE
)
}
\arguments{
\item{package}{An object with class ContentPackage.}

\item{report_name}{A character name to the exported file.}

\item{template_name}{A character with the name to the template report file in folder ./00_Template.}

\item{supp}{A boolean to specify if 'S' should be added before content number in export.}

\item{dataset}{A boolean to specify if export is a dataset type and not a report.}
}
\value{
A path to the report file.
}
\description{
Export package
}
\examples{
x <- create_content_package() |>
 add_to_package(
  create_content(
   create_content_backbone(
    'analysis of dataset',
    'T',
    function(dataset) {
     dataset |>
      flextable::flextable()
    }
   ),
   dataset = tibble::tibble(x = c(1, 2, 3))
  )
) |>
export_package('teste')
}
