% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_report.R
\name{export_report}
\alias{export_report}
\title{Export Report Package}
\usage{
export_report(
  package,
  template_name = "template_PT-BR.docx",
  supp = FALSE,
  add_toc = TRUE,
  spec = ""
)
}
\arguments{
\item{package}{An object of class `ContentPackage`.}

\item{template_name}{A character string specifying the name of the template report file located in the `./00_Template` folder. Defaults to "template_PT-BR.docx".}

\item{supp}{A boolean indicating whether 'S' should be added before the content number in the export. Defaults to FALSE.}

\item{add_toc}{A boolean indicating whether a table of contents should be added to the report. Defaults to TRUE.}

\item{spec}{A character string with path to the spec file.}
}
\value{
The path to the exported package.
}
\description{
This function exports a report package to a specified file format using a template. It allows customization of the template, and other export options such as adding a table of contents and supplementary content.
}
\examples{
\dontrun{
# Example usage:
# Create and export a content package
x <- create_content_package() |>
 add_to_package(
  create_content(
   create_content_backbone(
    'analysis of dataset',
    'T',
    function(dataset) {
     dataset |>
      flextable::flextable()
    }
   ),
   dataset = tibble::tibble(x = c(1, 2, 3))
  )
) |>
export_report('teste')
}
}
