% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hash.R
\name{gen_hash}
\alias{gen_hash}
\title{Generate a SHA1 Hash from a Log File}
\usage{
gen_hash(path)
}
\arguments{
\item{path}{A character string specifying the path to the file. The path must be a single character string and the file must exist.}
}
\value{
A character string representing the SHA1 hash of the file contents.
}
\description{
This function generates a SHA1 hash from the contents of a specified file.
}
\examples{
# Example usage:
\dontrun{
x <- gen_hash('./teste/table_teste_backbone.R')
}
}
