% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_str.R
\name{mock_str}
\alias{mock_str}
\title{Mask numbers in a string based on a pattern}
\usage{
mock_str(x, pattern)
}
\arguments{
\item{x}{A character vector with the strings to be masked.}

\item{pattern}{A character vector specifying the pattern(s) for each string.
Patterns can include:
\itemize{
  \item{"keep"}{ - keep the value as is}
  \item{"i(n)"}{ - integer with n digits}
  \item{"f(n,m)"}{ - float with n integer digits and m decimal digits, default decimal separator '.'}
  \item{"f(n,m,sep)"}{ - float with n integer digits, m decimal digits, and explicit decimal separator ('.' or ',')}
}}
}
\value{
A character vector with the numeric values masked according to the pattern.
}
\description{
Replaces numeric values in a string with a mock representation ("x")
according to a user-specified pattern. Non-numeric parts can be kept.
}
\examples{
mock_str("A H5N8 7,1 mas ok (N = 1000.0)", "keep+keep+i(1)+f(3,1)")
}
