% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_adam_attr.R
\name{set_adam_attr}
\alias{set_adam_attr}
\title{Set ADaM attributes for a given dataset and metadata file}
\usage{
set_adam_attr(dataset, path, name)
}
\arguments{
\item{dataset}{A tibble with the dataset to be compared.}

\item{path}{A character to specify the path of the metadata file in .xlsx format based on package template.}

\item{name}{A character to specify the name of dataset in the metadata file.}
}
\value{
A tibble.
}
\description{
Set ADaM attributes for a given dataset and metadata file
}
\examples{
x <- set_adam_attr(dataset, 'ADSL.xlsx')
}
