% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_adam_dataset.R
\name{simulate_adam_dataset}
\alias{simulate_adam_dataset}
\title{Simulate an analysis dataset based on created dataset}
\usage{
simulate_adam_dataset(dataset, n = 10)
}
\arguments{
\item{dataset}{A tibble created via create_adam_dataset function.}

\item{n}{A number with the number of rows to be simulated}
}
\value{
A tibble with the data simulation.
}
\description{
Simulate an analysis dataset based on created dataset
}
\examples{
x <- simulate_adam_dataset(create_adam_dataset('ADSL.xlsx'))
}
