% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_backbones.R
\name{source_backbones}
\alias{source_backbones}
\title{Source Backbones from Backbones Folder Within a Package}
\usage{
source_backbones(pkg_name)
}
\arguments{
\item{pkg_name}{A character string specifying the name of the package.}
}
\value{
A list of sourced backbones.
}
\description{
This function sources all backbones from the backbones folder within a specified package, returning the loaded backbones.
}
\examples{
\dontrun{
# Example usage:
# Source backbones from the package named 'teste2'
source_backbones('teste2')
}
}
