% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_adam_dataset.R
\name{validate_adam_dataset}
\alias{validate_adam_dataset}
\title{Validate ADaM Dataset Against Metadata File}
\usage{
validate_adam_dataset(dataset, print = TRUE)
}
\arguments{
\item{dataset}{A tibble containing the dataset to be validated.}

\item{print}{A boolean indicating if validation process should be printed..}
}
\value{
TRUE if the dataset is validated successfully; FALSE if there are mismatches.
}
\description{
This function validates a dataset by comparing it against a metadata file, checking for structural mismatches and attribute inconsistencies. It reports any issues found and suggests using `set_adam_attr()` to fix missing metadata paths.
}
\examples{
\dontrun{
# Example usage:
# Validate a dataset against the 'ADSL.xlsx' metadata file
x <- validate_adam_dataset(dataset)
}
}
