% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_package.R
\name{add_to_package}
\alias{add_to_package}
\title{Add Content to a Content Package}
\usage{
add_to_package(package, content)
}
\arguments{
\item{package}{An object of class `ContentPackage`. This represents the package to which content will be added.}

\item{content}{An object of class `Content`. This represents the content to be added to the package.}
}
\value{
An object of class `ContentPackage` with the new content added.
}
\description{
This function adds content to an existing content package. It requires that the package and content are of specific classes.
}
\examples{
\dontrun{
# Example usage:
# Create a content package and add content to it
x <- create_content_package() |>
 add_to_package(
  create_content(
   create_content_backbone(
    'analysis of dataset',
    'T',
    function(dataset) {
     dataset |>
      flextable::flextable()
    }
   ),
   dataset = tibble::tibble(x = c(1, 2, 3))
  )
)
}
}
