% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_backbone_structure.R
\name{create_backbone_structure}
\alias{create_backbone_structure}
\title{Create backbone structure folder}
\usage{
create_backbone_structure(backbone_name, backbone_type, pkg_name)
}
\arguments{
\item{backbone_name}{A character to specify the name of the backbone.}

\item{backbone_type}{A character to specify the type of the backbone. Please use 'T', 'F' or 'L'.}

\item{pkg_name}{A character to specify the name of the package.}
}
\value{
A path to the backbone structure folder
}
\description{
Create backbone structure folder
}
\examples{
create_backbone_structure('teste3', 'T', 'teste2')
}
