% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_content.R
\name{create_content}
\alias{create_content}
\title{Create content from content backbone}
\usage{
create_content(
  content_backbone,
  dataset,
  subtitle = NA_character_,
  population = NA_character_,
  section = NA_character_,
  fdim = list(width = 9, height = 5, dpi = 600),
  ...
)
}
\arguments{
\item{content_backbone}{An object with class ContentBackbone.}

\item{dataset}{A tibble dataset to apply the content backbone.}

\item{subtitle}{A character to specify the content subtitle (optional).}

\item{population}{A character to specify the content population (optional).}

\item{section}{A character to specify the content section (optional).}

\item{fdim}{A list of 3 elements 'width', 'height' and 'dpi', respectively, with 3 numeric values associated. Only used for type = 'F'.}

\item{...}{Extra arguments to pass to 'fun' in content backbone (optional).}
}
\value{
A content.
}
\description{
Create content from content backbone
}
\examples{
x <- create_content(
 create_content_backbone(
  'analysis of dataset',
  'T',
  function(dataset) {
   dataset |>
    flextable::flextable()
  }
 ),
 dataset = tibble::tibble(x = c(1, 2, 3))
)
}
