% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive2.R
\name{derive2}
\alias{derive2}
\title{Simplified Version of Derive Function}
\usage{
derive2(.data, var_target, var_source, by = rlang::exprs(USUBJID))
}
\arguments{
\item{.data}{A data frame (primary dataset) where the variable will be derived.}

\item{var_target}{Symbol. Name of the variable to be derived in `.data`.}

\item{var_source}{Symbol. Name of the variable to be used as a reference to derive in `.data`.}

\item{by}{Symbol. The subject identifier variable name common to `.data`.}
}
\value{
The original dataset `.data` with the variable `var_target` derived accordingly.
}
\description{
This function provides a simplified version of the derive function, intended for use when converting one variable into another with the same length of levels.
}
\examples{
\dontrun{
# Example usage:
# Derive variable SEXN from SEX in ADSL dataset
adsl <- derive2(adsl, SEX, SEXN, by = USUBJID)
}
}
