% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_display.R
\name{fig_display}
\alias{fig_display}
\title{Display figure}
\usage{
fig_display(fig, fdim = list(width = 9, height = 5, dpi = 600))
}
\arguments{
\item{fig}{A ggplot object.}

\item{fdim}{A list of 3 elements 'width', 'height' and 'dpi', respectively, with 3 numeric values associated. Only used for type = 'F'.}
}
\value{
A ggplot.
}
\description{
Display figure
}
\examples{
x <- fig_display(ggplot())
}
