% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_caller_contents.R
\name{run_caller_contents}
\alias{run_caller_contents}
\title{Run contents from caller_contents file}
\usage{
run_caller_contents(package, backbones, pkg_name)
}
\arguments{
\item{package}{An object with class ContentPackage.}

\item{backbones}{A list of backbones.}

\item{pkg_name}{A character with the package name.}
}
\value{
A package with content added.
}
\description{
Run contents from caller_contents file
}
\examples{
x <- run_caller_contents(package_init, backbones, 'name of the package')
}
