% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_adam_attr.R
\name{set_adam_attr}
\alias{set_adam_attr}
\title{Set ADaM Attributes for a Given Dataset and Metadata File}
\usage{
set_adam_attr(dataset, path, name)
}
\arguments{
\item{dataset}{A tibble containing the dataset to be updated with ADaM attributes.}

\item{path}{A character string specifying the path to the metadata file in .xlsx format, based on the package template.}

\item{name}{A character string specifying the name of the dataset in the metadata file.}
}
\value{
The dataset tibble with updated ADaM attributes.
}
\description{
This function sets ADaM attributes for a given dataset based on a specified metadata file. It ensures that the dataset's columns have the correct labels, data types, origins, and levels as defined in the metadata.
}
\examples{
\dontrun{
# Example usage:
# Set ADaM attributes for a dataset using metadata from 'ADSL.xlsx'
x <- set_adam_attr(dataset, 'ADSL.xlsx', 'ADSL')
}
}
