% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_backbones.R
\name{source_backbones}
\alias{source_backbones}
\title{Source backbones from backbones folder within a package}
\usage{
source_backbones(pkg_name)
}
\arguments{
\item{pkg_name}{A character to specify the name of the package.}
}
\value{
A path to the backbone structure folder
}
\description{
Source backbones from backbones folder within a package
}
\examples{
source_backbones('teste2')
}
