% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_adam_dataset.R
\name{validate_adam_dataset}
\alias{validate_adam_dataset}
\title{Create a report with comparisons against given dataset and metadata file and fix attributes issues}
\usage{
validate_adam_dataset(dataset, path)
}
\arguments{
\item{dataset}{A tibble with the dataset to be compared.}

\item{path}{A character to specify the path of the metadata file in .xlsx format based on package template.}
}
\value{
TRUE if dataset validated and FALSE if not.
}
\description{
Create a report with comparisons against given dataset and metadata file and fix attributes issues
}
\examples{
x <- validate_adam_dataset(dataset, 'ADSL.xlsx')
}
