% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_hash.R
\name{validate_hash}
\alias{validate_hash}
\title{Validate a hash SHA1 file given in log file.}
\usage{
validate_hash(hash, path)
}
\arguments{
\item{hash}{A character to specify the hash of the file.}

\item{path}{A character to specify the path of the file.}
}
\value{
TRUE or FALSE
}
\description{
Validate a hash SHA1 file given in log file.
}
\examples{
x <- validate_hash('abcdkfj28747', './teste/table_teste_backbone.R')
}
